Move configuration to its own module
