+++ test/unit/test_configuration.py
@@ -1,0 +1,76 @@
+# This file is part of PyOP2
+#
+# PyOP2 is Copyright (c) 2012, Imperial College London and
+# others. Please see the AUTHORS file in the main source directory for
+# a full list of copyright holders.  All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#     * Redistributions of source code must retain the above copyright
+#       notice, this list of conditions and the following disclaimer.
+#     * Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#     * The name of Imperial College London or that of other
+#       contributors may not be used to endorse or promote products
+#       derived from this software without specific prior written
+#       permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTERS
+# ''AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+# INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+# HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+# STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+# OF THE POSSIBILITY OF SUCH DAMAGE.
+
+"""Configuration unit tests."""
+
+import pytest
+from pyop2.configuration import Configuration
+from pyop2.exceptions import ConfigurationError
+
+
+class TestConfigurationAPI:
+    """Configuration API unit tests."""
+
+    def test_add_configuration_value(self):
+        """Defining an non default argument."""
+        c = Configuration()
+        c.reconfigure(foo='bar')
+        assert c['foo'] == 'bar'
+
+    def test_change_backend(self):
+        """backend option is read only."""
+        c = Configuration()
+        c.reconfigure(backend='cuda')
+        with pytest.raises(ConfigurationError):
+            c['backend'] = 'other'
+
+    def test_reconfigure_backend(self):
+        """backend option is read only."""
+        c = Configuration()
+        c.reconfigure(backend='cuda')
+        with pytest.raises(ConfigurationError):
+            c.reconfigure(backend='other')
+
+    @pytest.mark.parametrize(('key', 'val'), [('backend', 0),
+                                              ('debug', 'illegal'),
+                                              ('log_level', 1.5),
+                                              ('lazy_evaluation', 'illegal'),
+                                              ('lazy_max_trace_length', 'illegal'),
+                                              ('dump_gencode', 'illegal'),
+                                              ('dump_gencode_path', 0)])
+    def test_configuration_illegal_types(self, key, val):
+        """Illegal types for configuration values should raise
+        ConfigurationError."""
+        c = Configuration()
+        with pytest.raises(ConfigurationError):
+            c[key] = val
